fix: improve FTA score sim-aws-local-server.ts
