+++ src/serve/index.ts
@@ -1,2 +1,5 @@
 export { SimAwsHttp } from "./http/sim-aws-http.js";
-export { serveSimAws, SimAwsLocalServer } from "./http/sim-aws-local-server.js";
+export {
+  serveSimAws,
+  SimAwsLocalServer,
+} from "./http/local-server/sim-aws-local-server.js";
